delete :: 검증로직 함수로 분리 롤백

+++ bitgouel-api/src/main/kotlin/team/msg/domain/student/service/StudentActivityServiceImpl.kt
@@ -65,7 +65,16 @@ class StudentActivityServiceImpl(
      */
     @Transactional(rollbackFor = [Exception::class])
     override fun updateStudentActivity(id: UUID, request: UpdateStudentActivityRequest) {
-        val studentActivity = validateStudent(id)
+        val user = userUtil.queryCurrentUser()
+
+        val student = studentRepository.findByUser(user)
+            ?: throw StudentNotFoundException("학생을 찾을 수 없습니다. info : [ userId = ${user.id}, username = ${user.name} ]")
+
+        val studentActivity = studentActivityRepository.findByIdOrNull(id)
+            ?: throw StudentActivityNotFoundException("학생 활동을 찾을 수 없습니다. info : [ studentActivityId = $id ]")
+
+        if(student.id != studentActivity.student.id)
+            throw ForbiddenStudentActivityException("해당 학생 활동에 대한 권한이 없습니다. info : [ studentId = ${student.id} ]")
 
         val updatedStudentActivity = StudentActivity(
             id = studentActivity.id,
@@ -87,8 +96,17 @@ class StudentActivityServiceImpl(
      */
     @Transactional(rollbackFor = [Exception::class])
     override fun deleteStudentActivity(id: UUID) {
-        val studentActivity = validateStudent(id)
-        
+        val user = userUtil.queryCurrentUser()
+
+        val student = studentRepository.findByUser(user)
+            ?: throw StudentNotFoundException("학생을 찾을 수 없습니다. info : [ userId = ${user.id}, username = ${user.name} ]")
+
+        val studentActivity = studentActivityRepository.findByIdOrNull(id)
+            ?: throw StudentActivityNotFoundException("학생 활동을 찾을 수 없습니다. info : [ studentActivityId = $id ]")
+
+        if(student.id != studentActivity.student.id)
+            throw ForbiddenStudentActivityException("해당 학생 활동에 대한 권한이 없습니다. info : [ studentId = ${student.id} ]")
+
         studentActivityRepository.delete(studentActivity)
     }
 
@@ -117,16 +135,7 @@ class StudentActivityServiceImpl(
      * @param 학생활동을 검증하기 위한 id
      */
     private fun validateStudent(id: UUID): StudentActivity {
-        val user = userUtil.queryCurrentUser()
 
-        val student = studentRepository.findByUser(user)
-            ?: throw StudentNotFoundException("학생을 찾을 수 없습니다. info : [ userId = ${user.id}, username = ${user.name} ]")
-
-        val studentActivity = studentActivityRepository.findByIdOrNull(id)
-            ?: throw StudentActivityNotFoundException("학생 활동을 찾을 수 없습니다. info : [ studentActivityId = $id ]")
-
-        if(student.id != studentActivity.student.id)
-            throw ForbiddenStudentActivityException("해당 학생 활동에 대한 권한이 없습니다. info : [ studentId = ${student.id} ]")
 
         return studentActivity
     }
